refactor: move NowPlaying and PlaybackReport logic to separate worker files
