+++ core/scrobbler/play_tracker.go
@@ -363,132 +363,6 @@ func (p *playTracker) ReportPlayback(ctx context.Context, params ReportPlaybackP
 	return nil
 }
 
-func (p *playTracker) enqueueNowPlaying(ctx context.Context, playerId string, userId string, track *model.MediaFile, position int) {
-	p.npMu.Lock()
-	defer p.npMu.Unlock()
-	ctx = context.WithoutCancel(ctx) // Prevent cancellation from affecting background processing
-	p.npQueue[playerId] = nowPlayingEntry{
-		ctx:      ctx,
-		userId:   userId,
-		track:    track,
-		position: position,
-	}
-	p.sendNowPlayingSignal()
-}
-
-func (p *playTracker) sendNowPlayingSignal() {
-	// Don't block if the previous signal was not read yet
-	select {
-	case p.npSignal <- struct{}{}:
-	default:
-	}
-}
-
-func (p *playTracker) nowPlayingWorker() {
-	defer close(p.workerDone)
-	for {
-		select {
-		case <-p.shutdown:
-			return
-		case <-time.After(time.Second):
-		case <-p.npSignal:
-		}
-
-		p.npMu.Lock()
-		if len(p.npQueue) == 0 {
-			p.npMu.Unlock()
-			continue
-		}
-
-		// Keep a copy of the entries to process and clear the queue
-		entries := p.npQueue
-		p.npQueue = make(map[string]nowPlayingEntry)
-		p.npMu.Unlock()
-
-		// Process entries without holding lock
-		for _, entry := range entries {
-			p.dispatchNowPlaying(entry.ctx, entry.userId, entry.track, entry.position)
-		}
-	}
-}
-
-func (p *playTracker) enqueuePlaybackReport(ctx context.Context, userId string, info PlaybackSession) {
-	p.prMu.Lock()
-	defer p.prMu.Unlock()
-	ctx = context.WithoutCancel(ctx)
-	p.prQueue = append(p.prQueue, playbackReportEntry{
-		ctx:    ctx,
-		userId: userId,
-		info:   info,
-	})
-	p.sendPlaybackReportSignal()
-}
-
-func (p *playTracker) sendPlaybackReportSignal() {
-	select {
-	case p.prSignal <- struct{}{}:
-	default:
-	}
-}
-
-func (p *playTracker) playbackReportWorker() {
-	defer close(p.prWorkerDone)
-	for {
-		select {
-		case <-p.shutdown:
-			return
-		case <-p.prSignal:
-		}
-
-		p.prMu.Lock()
-		if len(p.prQueue) == 0 {
-			p.prMu.Unlock()
-			continue
-		}
-		entries := p.prQueue
-		p.prQueue = nil
-		p.prMu.Unlock()
-
-		allScrobblers := p.getActiveScrobblers()
-		for _, entry := range entries {
-			p.dispatchPlaybackReport(entry.ctx, entry.userId, entry.info, allScrobblers)
-		}
-	}
-}
-
-func (p *playTracker) dispatchPlaybackReport(ctx context.Context, userId string, info PlaybackSession, allScrobblers map[string]Scrobbler) {
-	for name, s := range allScrobblers {
-		if !s.IsAuthorized(ctx, userId) {
-			continue
-		}
-		log.Debug(ctx, "Sending PlaybackReport", "scrobbler", name, "track", info.MediaFile.Title, "state", info.State, "positionMs", info.PositionMs)
-		err := s.PlaybackReport(ctx, userId, info)
-		if err != nil {
-			log.Error(ctx, "Error sending PlaybackReport", "scrobbler", name, "track", info.MediaFile.Title, "state", info.State, err)
-			continue
-		}
-	}
-}
-
-func (p *playTracker) dispatchNowPlaying(ctx context.Context, userId string, t *model.MediaFile, position int) {
-	if t.Artist == consts.UnknownArtist {
-		log.Debug(ctx, "Ignoring external NowPlaying update for track with unknown artist", "track", t.Title, "artist", t.Artist)
-		return
-	}
-	allScrobblers := p.getActiveScrobblers()
-	for name, s := range allScrobblers {
-		if !s.IsAuthorized(ctx, userId) {
-			continue
-		}
-		log.Debug(ctx, "Sending NowPlaying update", "scrobbler", name, "track", t.Title, "artist", t.Artist, "position", position)
-		err := s.NowPlaying(ctx, userId, t, position)
-		if err != nil {
-			log.Error(ctx, "Error sending PlaybackSession", "scrobbler", name, "track", t.Title, "artist", t.Artist, err)
-			continue
-		}
-	}
-}
-
 func (p *playTracker) GetNowPlaying(_ context.Context) ([]PlaybackSession, error) {
 	res := p.playMap.Values()
 	slices.SortFunc(res, func(a, b PlaybackSession) int {
